+++ tests/dummy/app/controllers/ember-flexberry-dummy-suggestion-edit.js
@@ -45,6 +45,7 @@ export default BaseEditFormController.extend(EditFormControllerOperationsIndicat
             relationName: 'author',
             projection: 'ApplicationUserL',
             autocomplete: true,
+            lookupWindowCustomProperties: () => { return { filterButton: true, refreshButton: true }; },
           };
           break;