+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+* Fix `lookupWindowCustomProperties` for lookups in `flexberry-groupedit`.
 
 ## [0.9.0-beta.21] - 2017-09-25
 ### Fixed